Move auth service to /services
